Dev ramesh

+++ user_agent_classifier/src/ua_types.h
@@ -3,12 +3,12 @@
 
 typedef struct  _model_name_type { 
   char model[UA_MAX_LEN_MODEL_NAME+1];
-  float intercept;
+  double intercept;
 } MODEL_NAME_TYPE;
 
 typedef struct  _model_rec_type { 
   char word[UA_MAX_LEN_WORD+1];
-  float *coefficients;
+  double *coefficients;
 } MODEL_REC_TYPE;
 
 #endif
